pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.10.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
